# docs/04-api/openapi/frameflow-v1.yaml
openapi: 3.0.3
info:
  title: FrameFlow Select API
  version: 1.0.0
  description: FrameFlow Select 批量质检与 Top-K 优选平台。身份端点沿用 M01（/api/v1/auth、/api/v1/teams）；产品端点为手写 OpenAPI 权威契约，运行时以 springdoc 输出并经验证对齐。
...
paths:
...
  /api/v1/auth/register:
    post:
      tags:
      - auth
      summary: 用户注册
      description: 规范化邮箱已注册时返回 409 USER_EMAIL_CONFLICT。
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          description: 注册成功，返回用户信息
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 参数校验失败
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 邮箱已被注册
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/refresh:
    post:
      tags:
      - auth
      summary: 刷新 Token（轮换，旧 Refresh Token 立即失效）
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: 新 Token 对
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Refresh Token 无效/过期/重放
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/logout:
    post:
      tags:
      - auth
      summary: 登出，撤销 Refresh Token
      description: Refresh Token 是唯一凭据，不要求 Access Token；即使 Access Token 已过期，仍可撤销该设备 token family。
      operationId: logout
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '204':
          description: 登出成功
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
        '401':
          description: Refresh Token 无效/过期/重放
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: 登录，返回 Token 对
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: 登录成功
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          description: 参数校验失败
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 邮箱或密码错误
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: 当前登录用户
      operationId: currentUser
      responses:
        '200':
          description: 当前用户信息
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: 未认证或 Token 无效/过期
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
          description: 稳定错误码
          example: AUTH_REQUIRED
        message:
          type: string
          description: 人类可读的错误说明
        requestId:
          type: string
          format: uuid
          description: 必须与响应 X-Request-Id header 完全相同
        traceId:
          type: string
          description: 可观测环境可选
      required:
      - code
      - message
      - requestId
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: 邮箱（小写规范化）
          maxLength: 255
          minLength: 0
        password:
          type: string
          format: password
          description: 密码（BCrypt 存储，永不出现在响应）
          maxLength: 72
          minLength: 8
        displayName:
          type: string
          description: 显示名称
          maxLength: 100
          minLength: 1
      required:
      - displayName
      - email
      - password
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        displayName:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - DISABLED
          - PENDING
        lastLoginAt:
          type: string
          format: date-time
      required:
      - displayName
      - email
      - id
      - status
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          description: 32 字节安全随机值的 Base64URL 无 padding 编码
          maxLength: 43
          minLength: 43
      required:
      - refreshToken
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
          description: RS256 JWT，TTL 15 分钟
        refreshToken:
          type: string
          description: 32 字节安全随机的不透明 Base64URL 字符串；轮换且 TTL 30 天
          maxLength: 43
          minLength: 43
        expiresIn:
          type: integer
          format: int64
          description: Access Token 有效期（秒）
          example: 900
        tokenType:
          type: string
          enum:
          - Bearer
      required:
      - accessToken
      - expiresIn
      - refreshToken
      - tokenType
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: 邮箱（登录前先去除首尾空白并转小写）
        password:
          type: string
          format: password
          description: 密码
      required:
      - email
      - password
...
  headers:
    RequestId:
      description: 服务器为本次 HTTP 请求生成的请求 ID
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT